requirements.txt
crewai
crewai_tools
load_dotenv
langchain-groq
langchain-huggingface
langchain_google_genai
langchain